Default to blue
Defaulting to black seems like the sensible choice at first glance. However, it forces users to move the grayscale slider between being able to choose a color, which is janky and could be confusing.

Ideally we would remember the last used color, but for now this switch should eliminate quite a bit of confusion.